requirements.txt
requests
pandas
sqlalchemy
psycopg2-binary
pyyaml
python-dotenv
numpy